refactor: extract shared inspection assets and cleanup duplicate strings

+++ core/ui/src/commonMain/kotlin/dev/fanfly/wingslog/core/ui/common/datetime/DateTimeUtils.kt
@@ -1,15 +1,15 @@
 package dev.fanfly.wingslog.core.ui.common.datetime
 
+import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.format.char
 import kotlinx.datetime.toLocalDateTime
-import kotlin.time.Instant
 import com.squareup.wire.Instant as WireInstant
 
 fun WireInstant.toLocalDate(): LocalDate {
   // Convert WireInstant (epoch seconds and nanos) to LocalDate
-  val instant = Instant.fromEpochSeconds(this.getEpochSecond(), this.getNano())
+  val instant = Instant.fromEpochSeconds(this.getEpochSecond(), this.getNano().toLong())
   return instant.toLocalDateTime(TimeZone.UTC).date
 }
